!230 修复通知公告中存在的问题 * 修复撤回消息 * Merge branch 'master' of gitee.com:zhangjun1024/Admin.NET * 🏌️修复任何人都能管理其他人发布的通知通告的bug

+++ backend/Furion.Extras.Admin.NET/Service/Notice/SysNoticeService.cs
@@ -51,6 +51,8 @@ namespace Furion.Extras.Admin.NET.Service.Notice
                                                                       EF.Functions.Like(u.Content, $"%{input.SearchValue.Trim()}%"))
                                              .Where(input.Type > 0, u => u.Type == input.Type)
                                              .Where(u => u.Status != NoticeStatus.DELETED)
+                                             //通知公告管理应只有发布人可以管理自己发布的，其他人只能在已收公告中看到
+                                             .Where(u=>u.PublicUserId == _userManager.UserId)
                                              .ToADPagedListAsync(input.PageNo, input.PageSize);
             return notices;
         }
@@ -72,11 +74,21 @@ namespace Furion.Extras.Admin.NET.Service.Notice
             if (input.Status == NoticeStatus.PUBLIC)
                 notice.PublicTime = DateTimeOffset.Now;
             var newItem = await notice.InsertNowAsync();
-
-            // 通知到的人
-            var noticeUserIdList = input.NoticeUserIdList;
-            var noticeUserStatus = NoticeUserStatus.UNREAD;
-            await _sysNoticeUserService.Add(newItem.Entity.Id, noticeUserIdList, noticeUserStatus);
+            //如果是草稿则将状态标记为暂不通知，下次更改状态后再修改为未更改状态
+            if (notice.Status!=NoticeStatus.DRAFT)
+            {
+                // 通知到的人
+                var noticeUserIdList = input.NoticeUserIdList;
+                var noticeUserStatus = NoticeUserStatus.UNREAD;
+                await _sysNoticeUserService.Add(newItem.Entity.Id, noticeUserIdList, noticeUserStatus);
+            }
+            else
+            {
+                var noticeUserIdList = input.NoticeUserIdList;
+                var noticeUserStatus = NoticeUserStatus.NONOTICE;
+                await _sysNoticeUserService.Add(newItem.Entity.Id, noticeUserIdList, noticeUserStatus);
+            }
+           
         }
 
         /// <summary>
@@ -88,10 +100,10 @@ namespace Furion.Extras.Admin.NET.Service.Notice
         public async Task DeleteNotice(DeleteNoticeInput input)
         {
             var notice = await _sysNoticeRep.FirstOrDefaultAsync(u => u.Id == input.Id);
-            if (notice.Status != NoticeStatus.DRAFT && notice.Status != NoticeStatus.CANCEL) // 只能删除草稿和撤回的
+            if (notice.Status != NoticeStatus.DRAFT && notice.Status != NoticeStatus.CANCEL ) // 只能删除草稿和撤回的公告
                 throw Oops.Oh(ErrorCode.D7001);
-
-            //notice.Status = (int)NoticeStatus.DELETED;
+            if (notice.PublicUserId != _userManager.UserId)
+                throw Oops.Oh(ErrorCode.D7003);
             await notice.DeleteAsync();
         }
 
@@ -109,6 +121,10 @@ namespace Furion.Extras.Admin.NET.Service.Notice
             //  非草稿状态
             if (input.Status != NoticeStatus.DRAFT)
                 throw Oops.Oh(ErrorCode.D7002);
+            // 如果发布者非本人则不能修改
+            SysNotice noticeInDb = await _sysNoticeRep.DetachedEntities.Where(u=>u.Id==input.Id).FirstOrDefaultAsync();
+            if (noticeInDb.PublicUserId != _userManager.UserId)
+                throw Oops.Oh(ErrorCode.D7003);
 
             var notice = input.Adapt<SysNotice>();
             if (input.Status == NoticeStatus.PUBLIC)
@@ -117,11 +133,16 @@ namespace Furion.Extras.Admin.NET.Service.Notice
                 await UpdatePublicInfo(notice);
             }
             await notice.UpdateAsync();
-
-            // 通知到的人
-            var noticeUserIdList = input.NoticeUserIdList;
-            var noticeUserStatus = NoticeUserStatus.UNREAD;
-            await _sysNoticeUserService.Update(input.Id, noticeUserIdList, noticeUserStatus);
+            //如果修改后的状态为非草稿状态则通知
+            if (notice.Status != NoticeStatus.DRAFT)
+            {
+                // 通知到的人
+                var noticeUserIdList = input.NoticeUserIdList;
+                var noticeUserStatus = NoticeUserStatus.UNREAD;
+                await _sysNoticeUserService.Update(input.Id, noticeUserIdList, noticeUserStatus);
+            }
+          
+            
         }
 
         /// <summary>
@@ -156,8 +177,12 @@ namespace Furion.Extras.Admin.NET.Service.Notice
             var noticeResult = notice.Adapt<NoticeDetailOutput>();
             noticeResult.NoticeUserIdList = noticeUserIdList;
             noticeResult.NoticeUserReadInfoList = noticeUserReadInfoList;
+            if(noticeResult.Status==(int)NoticeStatus.CANCEL)
+            {
+                noticeResult.Content = "<h1 style=\"text-align: center; \">该内容已被发布者撤回</h1>";
+            }
             // 如果该条通知公告为已发布，则将当前用户的该条通知公告设置为已读
-            if (notice.Status == NoticeStatus.PUBLIC)
+            if (notice.Status == NoticeStatus.PUBLIC || notice.Status==NoticeStatus.CANCEL)
                 await _sysNoticeUserService.Read(notice.Id, _userManager.UserId, NoticeUserStatus.READ);
             return noticeResult;
         }
@@ -174,18 +199,32 @@ namespace Furion.Extras.Admin.NET.Service.Notice
             if (input.Status != NoticeStatus.CANCEL && input.Status != NoticeStatus.DELETED && input.Status != NoticeStatus.PUBLIC)
                 throw Oops.Oh(ErrorCode.D7000);
 
-            var notice = await _sysNoticeRep.FirstOrDefaultAsync(u => u.Id == input.Id);
-            notice.Status = input.Status;
+            var noticeuser = await _sysNoticeUserRep.DetachedEntities.Where(u => u.NoticeId == input.Id).Select(u=>u.UserId).ToListAsync();
 
+            var notice = await _sysNoticeRep.FirstOrDefaultAsync(u => u.Id == input.Id);
+            if (notice.PublicUserId != _userManager.UserId)
+            {
+                throw Oops.Oh(ErrorCode.D7003);
+            }
+            notice.Status = input.Status;
             if (input.Status == NoticeStatus.CANCEL)
             {
                 notice.CancelTime = DateTimeOffset.Now;
             }
             else if (input.Status == NoticeStatus.PUBLIC)
             {
+                
                 notice.PublicTime = DateTimeOffset.Now;
             }
             await notice.UpdateAsync();
+            if (notice.Status != NoticeStatus.DRAFT)
+            {
+                // 通知到的人
+                var noticeUserIdList = noticeuser;
+                var noticeUserStatus = NoticeUserStatus.UNREAD;
+                await _sysNoticeUserService.Update(input.Id, noticeUserIdList, noticeUserStatus);
+            }
+
         }
 
         /// <summary>
@@ -202,7 +241,7 @@ namespace Furion.Extras.Admin.NET.Service.Notice
                                              .Where(searchValue, u => EF.Functions.Like(u.u.Title, $"%{input.SearchValue.Trim()}%") ||
                                                                       EF.Functions.Like(u.u.Content, $"%{input.SearchValue.Trim()}%"))
                                              .Where(input.Type > 0, u => u.u.Type == input.Type)
-                                             .Where(u => u.u.Status != NoticeStatus.DELETED)
+                                             .Where(u => u.u.Status != NoticeStatus.DELETED && u.u.Status!=NoticeStatus.DRAFT)
                                              .Select(u => u.u.Adapt(u.e.Adapt<NoticeReceiveOutput>()))
                                              .ToADPagedListAsync(input.PageNo, input.PageSize);
             return notices;
